+++ application/src/main/resources/application.yml
@@ -69,7 +69,6 @@ security:
 # Attribute Based Access Control
 abac:
   enabled: false
-  audit-queue-class: org.openehealth.ipf.commons.audit.queue.AsynchronousAuditMessageQueue
   # Server URL incl. trailing "/"!
   server: http://localhost:3001/rest/v1/policy/execute/name/
   # Definition of the JWT claim which contains the organization ID.
@@ -201,7 +200,9 @@ management:
 # Audit Properties
 ipf:
   atna:
-    audit-enabled: false
+#    audit-enabled: false
+    audit-enabled: true
+    audit-queue-class: org.openehealth.ipf.commons.audit.queue.AsynchronousAuditMessageQueue
 
 # External Terminology Validation Properties
 validation:
